+++ src/esengine/renderer/plugins/UIElementPlugin.cpp
@@ -154,9 +154,9 @@ void UIElementPlugin::emitQuad(
     u32 vOff = buffers.appendVertices(LayoutId::Batch, verts, sizeof(verts));
     u32 baseVertex = vOff / sizeof(BatchVertex);
 
-    u16 indices[6];
+    u32 indices[6];
     for (u32 i = 0; i < 6; ++i) {
-        indices[i] = static_cast<u16>(baseVertex + QUAD_INDICES[i]);
+        indices[i] = static_cast<u32>(baseVertex + QUAD_INDICES[i]);
     }
     u32 iOff = buffers.appendIndices(LayoutId::Batch, indices, 6);
 
@@ -255,9 +255,9 @@ void UIElementPlugin::emitNineSlice(
             u32 vOff = buffers.appendVertices(LayoutId::Batch, verts, sizeof(verts));
             u32 baseVert = vOff / sizeof(BatchVertex);
 
-            u16 indices[6];
+            u32 indices[6];
             for (u32 i = 0; i < 6; ++i) {
-                indices[i] = static_cast<u16>(baseVert + QUAD_INDICES[i]);
+                indices[i] = static_cast<u32>(baseVert + QUAD_INDICES[i]);
             }
             u32 iOff = buffers.appendIndices(LayoutId::Batch, indices, 6);
